✨ :: DataSourceImpl 에서 DTO 변경값에 따라 값 변경

+++ app/src/main/java/com/g3c1/oasis_android/feature_menu/data/dto/OrderCheckDTO.kt
@@ -2,6 +2,6 @@ package com.g3c1.oasis_android.feature_menu.data.dto
 
 
 data class OrderCheckDTO (
-    val seatId: Int,
-    val serialNumber: Int
+    val seatId: Long,
+    val serialNumber: Long
 )
